+++ papyrus/basepage.py
@@ -1,36 +1,64 @@
+import frontmatter
 import pathlib
 import datetime
 import os
+import re
 
 class BasePage():
-    def __init__(self, input_path, meta, content):
+    def __init__(self, input_path):
         self.input_path = input_path
-        self.meta = meta
-        self.content = content
-
+        self.content_text = None
+        self.content_html = None
+        self.excerpt_text = None
+        self.excerpt_html = None
         self.output_path = None
         self.layout = None
         self.url = None
         self.props = None
         self.title = ""
         self.date = None
-        self.wordcount = 0
         self.tags = []
-        self.content_text = None
-        self.content_html = None
-        self.excerpt_text = None
-        self.excerpt_html = None
 
-    def build(self, pandoc, config):
-        self.title, content = self.extract_title(self.content)
-        self.excerpt_text, self.content_text = \
-            self.get_excerpt_content(content, config['EXCERPT_SEPARATOR'])
+        self.__update()
 
+    def __update(self):
+        with self.input_path.open() as infile:
+            content = infile.read()
+        # parse and remove front matter
+        front_meta, content = frontmatter.parse(content)
+        # parse and remove back matter
+        back_meta, content = self.parse_backmatter(content)
+
+        self.meta = front_meta | back_meta
+        self.title, self.content = self.extract_title(content)
+        if not self.title:
+            self.title = self.meta['title'] if 'title' in self.meta else None
         self.layout = self.meta['layout'] if 'layout' in self.meta else 'post'
         self.date = self.get_date(self.meta, self.input_path)
         self.url = self.get_url(self.meta)
         self.props = self.meta['props'].split(" ") if "props" in self.meta else []
+        self.tags = self.meta['tags'].split(" ") if ('tags' in self.meta and self.meta['tags']) else []
+
+    def build(self, pandoc, config):
+        self.__update()
+
+        self.excerpt_text, self.content_text = \
+            self.get_excerpt_content(self.content, config['EXCERPT_SEPARATOR'])
         self.output_path = self.get_output_path(config['DIR_PUBLISH'])
+
+    @staticmethod
+    def parse_backmatter(content):
+        back_meta = {}
+        ii = []
+        for mo in re.finditer(r"^---\s*$", content, re.MULTILINE):
+            ii.append(mo.start())
+
+        if (len(ii)) == 2 and len(content) == ii[1] + 3:
+            back_matter = content[ii[0]:]
+            back_meta, _ = frontmatter.parse(back_matter)
+            content = content[:ii[0]].strip()
+
+        return back_meta, content
 
     @staticmethod
     def extract_title(content):
@@ -40,7 +68,12 @@ class BasePage():
             content = content[ii_first_newline:].strip()
             return first_line.replace("#", "").strip(), content
 
-        return "", content
+        return None, content
+
+    def is_draft(self):
+        if 'draft' in self.props:
+            return True
+        return False
 
     def is_public(self):
         if not self.props:
@@ -113,10 +146,9 @@ class BasePage():
         raise NotImplementedError()
 
     def write(self, jinja_env, site_meta):
-        if not self.is_current():
-            template_out = self.render(jinja_env, site_meta, self.get_page_metadata())
-            self.output_path.parent.mkdir(parents=True, exist_ok=True)
-            self.output_path.write_text(template_out)
+        template_out = self.render(jinja_env, site_meta, self.get_page_metadata())
+        self.output_path.parent.mkdir(parents=True, exist_ok=True)
+        self.output_path.write_text(template_out)
 
     def get_page_metadata(self):
         page = {