+++ papyrus/config.py
@@ -2,7 +2,6 @@ import pathlib
 
 DEFAULT_CONFIG_FILE = "_site_config.py"
 BOOTSTRAP_PATH = "bootstrap"
-TEMPLATES_PATH = "templates"
 
 def read_config():
     config_path = pathlib.Path(DEFAULT_CONFIG_FILE)